add endpoints + test profiles

+++ app/database/src/main/kotlin/com/codingmill/bookapi/database/repository/BookRepository.kt
@@ -3,7 +3,7 @@ package com.codingmill.bookapi.database.repository
 import com.codingmill.bookapi.database.entity.BookEntity
 import org.springframework.data.repository.CrudRepository
 
-interface BookRepository : CrudRepository<BookEntity, Int> {
+interface BookRepository : CrudRepository<BookEntity, Long> {
